Implement bookloupe
This tool is not identical in behavior to the standalone tool
used in GG1, but it does do essentially the same tests, and
reports them in a similar way.

There will be reports from this that the GG1 tool does not
report, and vice versa.

The aim is to provide a useful tool for PPers.

Future work will probably include more on curly quotes, and
other features of "recent" DP books.