Alternative to #668 (less-smart undo)
See #668 for main attempt to solve this issue.
In this commit, if you do an undo/redo, the list of
messages in the illo/SN checkers is removed until you
manually click Re-run.

This is just to see what it's like, and what you think is
best. It would probably be possible (though more
complicated) to gray out the text instead of removing it
and intercept user attempts to click on the messages and
either do nothing or put up a message saying "you need to
re-run the tool".